アービトラージをシミュレートするコードを追加
